move passwd to unix
