+++ alacritty/src/child_pty/unix/mod.rs
@@ -21,9 +21,14 @@ use nix::{
     unistd::setsid,
 };
 
-
 use die::die;
 
+#[cfg(not(target_os = "macos"))]
+use std::env;
+use std::ffi::CStr;
+use std::mem::MaybeUninit;
+
+use std::ptr;
 
 
 pub const PTY_BUFFER_SIZE: usize = 0x500;
@@ -32,6 +37,63 @@ mod ioctl {
     nix::ioctl_none_bad!(set_controlling, libc::TIOCSCTTY);
     nix::ioctl_write_ptr_bad!(win_resize, libc::TIOCSWINSZ, libc::winsize);
 }
+
+
+
+#[derive(Clone, Debug, Eq, PartialEq, Hash)]
+pub struct Passwd<'a> {
+    pub name: &'a str,
+    pub passwd: &'a str,
+    pub uid: libc::uid_t,
+    pub gid: libc::gid_t,
+    pub gecos: &'a str,
+    pub dir: &'a str,
+    pub shell: &'a str,
+}
+
+/// Return a Passwd struct with pointers into the provided buf.
+///
+/// # Unsafety
+///
+/// If `buf` is changed while `Passwd` is alive, bad thing will almost certainly happen.
+pub fn get_pw_entry(buf: &mut [i8; 1024]) -> Passwd<'_> {
+    // Create zeroed passwd struct.
+    let mut entry: MaybeUninit<libc::passwd> = MaybeUninit::uninit();
+
+    let mut res: *mut libc::passwd = ptr::null_mut();
+
+    // Try and read the pw file.
+    let uid = unsafe { libc::getuid() };
+    let status = unsafe {
+        libc::getpwuid_r(uid, entry.as_mut_ptr(), buf.as_mut_ptr() as *mut _, buf.len(), &mut res)
+    };
+    let entry = unsafe { entry.assume_init() };
+
+    if status < 0 {
+        // die!("getpwuid_r failed");
+    }
+
+    if res.is_null() {
+        // die!("pw not found");
+    }
+
+    // Sanity check.
+    assert_eq!(entry.pw_uid, uid);
+
+    // Build a borrowed Passwd struct.
+    Passwd {
+        name: unsafe { CStr::from_ptr(entry.pw_name).to_str().unwrap() },
+        passwd: unsafe { CStr::from_ptr(entry.pw_passwd).to_str().unwrap() },
+        uid: entry.pw_uid,
+        gid: entry.pw_gid,
+        gecos: unsafe { CStr::from_ptr(entry.pw_gecos).to_str().unwrap() },
+        dir: unsafe { CStr::from_ptr(entry.pw_dir).to_str().unwrap() },
+        shell: unsafe { CStr::from_ptr(entry.pw_shell).to_str().unwrap() },
+    }
+}
+
+
+
 
 
 
@@ -134,7 +196,7 @@ impl Pty {
         }
 
         let mut buf = [0; 1024];
-        let pw = crate::passwd::get_pw_entry(&mut buf);
+        let pw = get_pw_entry(&mut buf);
 
 
         let slave = pty.slave.clone();
