+++ alacritty/src/panic.rs
@@ -1,4 +1,4 @@
-
+use crate::child_pty::windows::win32_string;
 
 // Install a panic handler that renders the panic in a classical Windows error
 // dialog box as well as writes the panic to STDERR.
